Makefile: provide quite and verbose build
Do not print the whole commmands while building all the source code,
only print the file that is compiled and potential warning/errors.

If the source code is compiled using `make V=1` all the commands will be
printed (verbose mode on).

This is similar to what Linux's Makefile is doing.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>